Fix KSZap: rename, add configMaster, add/remove factory from whitelist
